refactor: Extract provider class
