+++ lib/screens/add_task_screen.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
-import 'package:todoey_flutter/main.dart';
+import 'package:todoey_flutter/models/task_provider.dart';
 
 class AddTaskScreen extends StatelessWidget {
   @override
